Added docco for annoted source documentation.  - After npm start docco documentation is available at localhost:8080/docs  - Cleaned up Rebound Runtime and Rebound Data files, moving on down the line to Model next.  - Broke Runtime out to its own top level file. I don't want any 'exposed' code - something that makes browser accessable objects - in the packages sub folder  - Consolidated the Component object. Hydrate and register are now methods on the Component object. Env and register files removed as they are un needed.  - We need to consolidate precompiler and compiler.  - I removed rebound.component.js. The only difference between that and the runtime is the 250 lines of code that make up the router, which is completly optionsal and is not started if no config option is passed.
