+++ .gitignore
@@ -24,6 +24,7 @@ tmp
 amd
 dist
 test/demo/templates
+docs
 
 # Dependency directory
 # Deployed apps should consider commenting this line out:
