+++ packages/rebound-component/lib/component.js
@@ -1,18 +1,55 @@
-import $ from "rebound-component/utils";
-import env from "rebound-component/env";
-import Context from "rebound-component/context";
+// Rebound Component
+// ----------------
 
-// If Rebound Runtime has already been run, throw error
-if(window.Rebound && window.Rebound.Component){
-  throw 'Rebound is already loaded on the page!';
-}
+import DOMHelper from "morph/dom-helper";
+import hooks from "rebound-component/hooks";
+import helpers from "rebound-component/helpers";
+import $ from "rebound-component/utils";
+import { Model } from "rebound-data/rebound-data";
+
 // If Backbone hasn't been started yet, throw error
-if(!window.Backbone){
-  throw "Backbone must be on the page for Rebound to load.";
+if(!window.Backbone) throw "Backbone must be on the page for Rebound to load.";
+
+// Returns true if `str` starts with `test`
+function startsWith(str, test){
+  if(str === test) return true;
+  return str.substring(0, test.length+1) === test+'.';
 }
+
+function renderCallback(){
+  var i = 0, len = this._toRender.length;
+  delete this._renderTimeout;
+  for(i=0;i<len;i++){
+    this._toRender.shift().call();
+  }
+}
+
+var env = {
+  helpers: helpers.helpers,
+  hooks: hooks
+};
+
+env.hydrate = function hydrate(spec, options){
+  // Return a wrapper function that will merge user provided helpers and hooks with our defaults
+  return function(data, options){
+    // Ensure we have a well-formed object as var options
+    var env = options || {},
+        contextElement = data.el || document.body;
+    env.helpers = env.helpers || {};
+    env.hooks = env.hooks || {};
+    env.dom = env.dom || new DOMHelper();
+
+    // Merge our default helpers and hooks with user provided helpers
+    env.helpers = _.defaults(env.helpers, helpers.helpers);
+    env.hooks = _.defaults(env.hooks, hooks);
+
+    // Call our func with merged helpers and hooks
+    return spec.render(data, env, contextElement);
+  };
+};
 
 // New Backbone Component
-var Component = Context.extend({
+var Component = Model.extend({
 
   isComponent: true,
 
@@ -24,7 +61,7 @@ var Component = Context.extend({
     this.changed = {};
     this.helpers = {};
     this.__parent__ = this.__root__ = this;
-    this.listen();
+    this.listenTo(this, 'all', this._onChange);
 
     // Take our parsed data and add it to our backbone data structure. Does a deep defaults set.
     // In the model, primatives (arrays, objects, etc) are converted to Backbone Objects
@@ -55,9 +92,10 @@ var Component = Context.extend({
     }
 
     // Take our precompiled template and hydrates it. When Rebound Compiler is included, can be a handlebars template string.
+    // TODO: Check if template is a string, and if the compiler exists on the page, and compile if needed
     if(!options.template && !this.template){ throw('Template must provided for ' + this.__name + ' component!'); }
     this.template = options.template || this.template;
-    this.template = (typeof options.template === 'string') ? Rebound.templates[this.template] : env.hydrate(this.template);
+    this.template = (typeof this.template === 'object') ? env.hydrate(this.template) : this.template;
 
 
     // Render our dom and place the dom in our custom element
@@ -107,13 +145,76 @@ var Component = Context.extend({
     // else{ this.set(attrName, newVal, {quiet: true}); }
   },
 
+
+  _onChange: function(type, model, collection, options){
+    var shortcircuit = { change: 1, sort: 1, request: 1, destroy: 1, sync: 1, error: 1, invalid: 1, route: 1, dirty: 1 };
+    if( shortcircuit[type] ) return;
+
+    var data, changed;
+    model || (model = {});
+    collection || (collection = {});
+    options || (options = {});
+    !collection.isData && (options = collection) && (collection = model);
+    this._toRender || (this._toRender = []);
+
+    if( (type === 'reset' && options.previousAttributes) || type.indexOf('change:') !== -1){
+      data = model;
+      changed = model.changedAttributes();
+    }
+    else if(type === 'add' || type === 'remove' || (type === 'reset' && options.previousModels)){
+      data = collection;
+      changed = {};
+      changed[data.__path()] = data;
+    }
+
+    if(!data || !changed) return;
+
+    var push = Array.prototype.push;
+    var context = this;
+    var basePath = data.__path();
+    var parts = $.splitPath(basePath);
+    var key, obsPath, path, observers;
+
+    // For each changed key, walk down the data tree from the root to the data
+    // element that triggered the event and add all relevent callbacks to this
+    // object's _toRender queue.
+    do{
+      for(key in changed){
+        path = (basePath + (basePath && '.') + key).replace(context.__path(), '').replace(/\[[^\]]+\]/g, ".@each").replace(/^\./, '');
+        for(obsPath in context.__observers){
+          observers = context.__observers[obsPath];
+          if(startsWith(obsPath, path) || startsWith(path, obsPath)){
+            // If this is a collection event, trigger everything, otherwise only trigger property change callbacks
+            if(data.isCollection) push.apply(this._toRender, observers.collection);
+            push.apply(this._toRender, observers.model);
+          }
+        }
+      }
+    } while(context !== data && (context = context.get(parts.shift())))
+
+    this._toRender = _.uniq(this._toRender);
+
+    // Queue our render callback to be called after the current call stack has been exhausted
+    window.clearTimeout(this._renderTimeout);
+    this._renderTimeout = window.setTimeout(_.bind(renderCallback, this), 0);
+  }
+
 });
 
 Component.extend= function(protoProps, staticProps) {
   var parent = this,
       child,
-      reservedMethods = {'trigger':1, 'constructor':1, 'get':1, 'set':1, 'has':1, 'extend':1, 'escape':1, 'unset':1, 'clear':1, 'cid':1, 'attributes':1, 'changed':1, 'toJSON':1, 'validationError':1, 'isValid':1, 'isNew':1, 'hasChanged':1, 'changedAttributes':1, 'previous':1, 'previousAttributes':1},
-      configProperties = {'routes':1, 'template':1, 'defaults':1, 'outlet':1, 'url':1, 'urlRoot':1, 'idAttribute':1, 'id':1, 'createdCallback':1, 'attachedCallback':1, 'detachedCallback':1};
+      reservedMethods = {
+        'trigger':1,    'constructor':1, 'get':1,               'set':1,             'has':1,
+        'extend':1,     'escape':1,      'unset':1,             'clear':1,           'cid':1,
+        'attributes':1, 'changed':1,     'toJSON':1,            'validationError':1, 'isValid':1,
+        'isNew':1,      'hasChanged':1,  'changedAttributes':1, 'previous':1,        'previousAttributes':1
+      },
+      configProperties = {
+        'routes':1,     'template':1,    'defaults':1, 'outlet':1,          'url':1,
+        'urlRoot':1,    'idAttribute':1, 'id':1,       'createdCallback':1, 'attachedCallback':1,
+        'detachedCallback':1
+      };
 
   protoProps.defaults = {};
 
@@ -157,5 +258,39 @@ Component.extend= function(protoProps, staticProps) {
   return child;
 };
 
+Component.register = function registerComponent(name, options) {
+  var script = options.prototype;
+  var template = options.template;
+  var style = options.style;
+
+  var component = this.extend(script, { __name: name });
+  var proto = Object.create(HTMLElement.prototype, {});
+
+  proto.createdCallback = function() {
+    this.__component__ = new component({
+      template: template,
+      outlet: this,
+      data: Rebound.seedData
+    });
+  };
+
+  proto.attachedCallback = function() {
+    script.attachedCallback && script.attachedCallback.call(this.__component__);
+  };
+
+  proto.detachedCallback = function() {
+    script.detachedCallback && script.detachedCallback.call(this.__component__);
+    this.__component__.deinitialize();
+  };
+
+  proto.attributeChangedCallback = function(attrName, oldVal, newVal) {
+    this.__component__._onAttributeChange(attrName, oldVal, newVal);
+    script.attributeChangedCallback && script.attributeChangedCallback.call(this.__component__, attrName, oldVal, newVal);
+  };
+
+  return document.registerElement(name, { prototype: proto });
+}
+
+_.bindAll(Component, 'register');
 
 export default Component;
