+++ packages/rebound-component/lib/helpers.js
@@ -1,3 +1,6 @@
+// Rebound Helpers
+// ----------------
+
 import LazyValue from "rebound-component/lazy-value";
 import $ from "rebound-component/utils";
