+++ packages/rebound-component/lib/hooks.js
@@ -1,7 +1,9 @@
+// Rebound Hooks
+// ----------------
+
 import LazyValue from "rebound-component/lazy-value";
 import $ from "rebound-component/utils";
 import helpers from "rebound-component/helpers";
-import Context from "rebound-component/context";
 
 var hooks = {},
     attributes = {  abbr: 1,      "accept-charset": 1,   accept: 1,      accesskey: 1,     action: 1,
@@ -146,14 +148,14 @@ var subtreeObserver = new MutationObserver(cleanSubtree);
 ********************************/
 
 hooks.get = function get(env, context, path){
-  context.blockParams || (context.blockParams = new Context());
+  context.blockParams || (context.blockParams = {});
   if(path === 'this'){ path = ''; }
   // context = (context.blockParams.has(path)) ? context.blockParams : context;
   return streamProperty(context, path);
 };
 
 hooks.set = function set(env, context, name, value){
-  context.blockParams || (context.blockParams = new Context());
+  context.blockParams || (context.blockParams = {});
   // context.blockParams.set(name, value);
 };
