+++ packages/rebound-data/lib/model.js
@@ -1,10 +1,8 @@
+// Rebound Model
+// ----------------
+
 import ComputedProperty from "rebound-data/computed-property";
 import $ from "rebound-component/utils";
-
-// If Rebound Runtime has already been run, throw error
-if(window.Rebound && window.Rebound.Model){ throw 'Rebound Model is already loaded on the page!'; }
-// If Backbone hasn't been started yet, throw error
-if(!window.Backbone){ throw "Backbone must be on the page for Rebound to load."; }
 
 function pathGenerator(parent, key){
   return function(){
