+++ wrap/end.runtime.frag
@@ -3,7 +3,7 @@
 //module value for 'main' here and return it as the
 //value to use for the public API for the built file.
 
-  return require(['rebound-runtime/rebound-runtime']);
+  return require(['runtime']);
 }));
 
 require.config({
